refactor: text-input use-enter-confirm
