+++ src/components/common/form/TextInput/TextInput.module.css
@@ -1,4 +1,4 @@
-@value main-yellow, border-gray, main-gray, warning-red, INPUT_HEIGHT, main-gray from "../variables.module.css";
+@value main-yellow, border-gray, main-gray, warning-red, INPUT_HEIGHT, main-gray from "../../variables.module.css";
 
 .wrapper {
   position: relative;
